Rename hash_map and dash_map modules to hashamp and dashmap respectively
The name reads better, and corresponds to the actual name of the crate.
