+++ src/state/keyed/hashmap.rs
